Improve benchmarking to run on each action and track results

+++ include/n3n/hexdump.h
@@ -6,8 +6,10 @@
 #ifndef HEXDUMP_H
 #define HEXDUMP_H
 
+#include <stdint.h>
 #include <stdio.h>
 
-void fhexdump (unsigned int display_addr, const void *in, int size, FILE *stream);
+
+void fhexdump (uint64_t display_addr, const void *in, int size, FILE *stream);
 
 #endif
